backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
requests
beautifulsoup4
python-whois